fix: rename PlayerAnimation to be more generic
